Introduction

Most of the functions work on a KPS3 object. For this, the variable s is used. Such a variable can be created with the lines:

using KiteModels, KitePodSimulator
const s = KPS3(KCU())

Input functions

KiteModels.set_v_reel_out — Function

set_v_reel_out(s::AKM, v_reel_out, t_0, period_time = 1.0 / s.set.sample_freq)

Setter for the reel-out speed. Must be called on every timestep (before each simulation). It also updates the tether length, therefore it must be called even if v_reel_out has not changed.

t_0 the start time of the next timestep relative to the start of the simulation [s]

KiteModels.set_depower_steering — Function

set_depower_steering(s::KPS3, depower, steering)

Setter for the depower and steering model inputs.

valid range for steering: -1.0 .. 1.0.

valid range for depower: 0 .. 1.0

This function sets the variables s.depower, s.steering and s.alpha_depower.

It takes the depower offset c0 and the dependency of the steering sensitivity from the depower settings into account.

KiteModels.set_v_wind_ground — Function

set_v_wind_ground(s::AKM, height, v_wind_gnd=s.set.v_wind, wind_dir=0.0)

Set the vector of the wind-velocity at the height of the kite. As parameter the height, the ground wind speed and the wind direction are needed. Must be called every at each timestep.

Output functions

KiteModels.unstretched_length — Function

unstretched_length(s::AKM)

Getter for the unstretched tether reel-out lenght (at zero force).

KiteModels.tether_length — Function

tether_length(s::AKM)

Calculate and return the real, stretched tether lenght.

KiteModels.winch_force — Function

winch_force(s::AKM)

Return the absolute value of the force at the winch as calculated during the last timestep.

KiteModels.spring_forces — Function

spring_forces(s::AKM)

Return an array of the scalar spring forces of all tether segements.

KiteModels.lift_drag — Function

lift_drag(s::AKM)

Return a tuple of the scalar lift and drag forces.

Example:

lift, drag = lift_drag(kps)

KiteModels.lift_over_drag — Function

lift_over_drag(s::AKM)

Return the lift-over-drag ratio.

KiteModels.v_wind_kite — Function

v_wind_kite(s::AKM)

Return the vector of the wind speed at the height of the kite.

Callback function for the DAE solver

KiteModels.residual! — Function

residual!(res, yd, y::MVector{S, SimFloat}, s::KPS3, time) where S

N-point tether model, one point kite at the top:
Inputs:
State vector y   = pos1, pos2, ..., posn, vel1, vel2, ..., veln
Derivative   yd  = vel1, vel2, ..., veln, acc1, acc2, ..., accn
Output:
Residual     res = res1, res2 = pos1,  ..., vel1, ...

Additional parameters:
s: Struct with work variables, type KPS3
S: The dimension of the state vector

The number of the point masses of the model N = S/6, the state of each point is represented by two 3 element vectors.

Environment

KiteModels.calc_rho — Function

calc_rho(s, height)

Calculate the air densisity as function of height.

KiteModels.calc_wind_factor — Function

calc_wind_factor(s, height, profile_law=s.set.profile_law)

Calculate the relative wind speed at a given height and reference height.

Helper functions

KiteModels.clear — Function

clear(s::KPS3)

Initialize the kite power model.

KiteModels.find_steady_state — Function

find_steady_state(s::KPS3, prn=false)

Find an initial equilibrium, based on the inital parameters l_tether, elevation and v_reel_out.

KiteModels.calc_drag — Function

calc_drag(s::KPS3, v_segment, unit_vector, rho, last_tether_drag, v_app_perp, area)

Calculate the drag of one tether segment, result stored in parameter last_tether_drag. Return the norm of the apparent wind velocity.

KiteModels.calc_set_cl_cd — Function

calc_set_cl_cd(s, vec_c, v_app)

Calculate the lift over drag ratio as a function of the direction vector of the last tether segment, the current depower setting and the apparent wind speed. Set the calculated CL and CD values in the struct s.