Introduction

Most of the functions work on a KPS3 object. For this, the variable s is used. Such a variable can be created with the lines:

using KiteUtils, KitePodSimulator
const s = KPS3(KCU())

Input functions

KiteModels.set_v_reel_out — Function

set_v_reel_out(s::AKM, v_reel_out, t_0, period_time = 1.0 / s.set.sample_freq)

Setter for the reel-out speed. Must be called on every timestep (before each simulation). It also updates the tether length, therefore it must be called even if v_reelout has not changed.

t_0 the start time of the next timestep relative to the start of the simulation [s]

KiteModels.set_l_tether — Function

set_l_tether(s::AKM, l_tether)

Setter for the tether reel-out lenght (at zero force). During real-time simulations use the function set_v_reel_out instead.

KiteModels.set_depower_steering — Function

set_depower_steering(s::KPS3, depower, steering)

Setter for the depower and steering model inputs.

valid range for steering: -1.0 .. 1.0.

valid range for depower: 0 .. 1.0

This function sets the variables s.depower, s.steering and s.alpha_depower.

It takes the depower offset c0 and the dependency of the steering sensitivity from the depower settings into account.

Output functions

KiteModels.get_l_tether — Function

get_l_tether(s::AKM)

Getter for the tether reel-out lenght (at zero force).

KiteModels.get_force — Function

get_force(s::AKM)

Return the absolute value of the force at the winch as calculated during the last timestep.

KiteModels.get_spring_forces — Function

get_spring_forces(s::AKM, pos)

Returns an array of the scalar spring forces of all tether segements.

Input: The vector pos of the positions of the point masses that belong to the tether.

KiteModels.get_lift_drag — Function

get_lift_drag(s::AKM)

Returns a tuple of the scalar lift and drag forces.

Example:

lift, drag = get_lift_drag(kps)

KiteModels.get_lod — Function

get_lod(s::AKM)

Returns the lift-over-drag ratio.

Callback function for the DAE solver

KiteModels.residual! — Function

residual!(res, yd, y::MVector{S, SimFloat}, s::KPS3, time) where S

N-point tether model, one point kite at the top:
Inputs:
State vector y   = pos1, pos2, ..., posn, vel1, vel2, ..., veln
Derivative   yd  = vel1, vel2, ..., veln, acc1, acc2, ..., accn
Output:
Residual     res = res1, res2 = pos1,  ..., vel1, ...

Struct with work variables: s of type KPS3
The parameter S is the dimension of the state vector.
N = S/6, each point is represented by two 3 element vectors.

Conversion functions