Introduction

Most of the functions work on a KPS3 or KPS4 object. For this, the variable s is used. Such a variable can be created with the lines:

using KiteModels, KitePodModels
const s = KPS3(KCU())

Or, if you want to use the 4 point kite model:

using KiteModels, KitePodModels
const s = KPS4(KCU())

Input functions

KiteModels.set_v_reel_out — Function

set_v_reel_out(s::AKM, v_reel_out, t_0, period_time = 1.0 / s.set.sample_freq)

Setter for the reel-out speed. Must be called on every timestep (before each simulation). It also updates the tether length, therefore it must be called even if v_reel_out has not changed.

t_0 the start time of the next timestep relative to the start of the simulation [s]

KiteModels.set_depower_steering — Function

set_depower_steering(s::KPS3, depower, steering)

Setter for the depower and steering model inputs.

valid range for steering: -1.0 .. 1.0.

valid range for depower: 0 .. 1.0

This function sets the variables s.depower, s.steering and s.alpha_depower.

It takes the depower offset c0 and the dependency of the steering sensitivity from the depower settings into account.

KiteModels.set_v_wind_ground — Function

set_v_wind_ground(s::AKM, height, v_wind_gnd=s.set.v_wind, wind_dir=0.0)

Set the vector of the wind-velocity at the height of the kite. As parameter the height, the ground wind speed [m/s] and the wind direction [radians] are needed. Must be called every at each timestep.

Output functions

KiteModels.unstretched_length — Function

unstretched_length(s::AKM)

Getter for the unstretched tether reel-out lenght (at zero force).

KiteModels.tether_length — Function

tether_length(s::AKM)

Calculate and return the real, stretched tether lenght.

KiteModels.winch_force — Function

winch_force(s::AKM)

Return the absolute value of the force at the winch as calculated during the last timestep.

KiteModels.spring_forces — Function

spring_forces(s::AKM)

Return an array of the scalar spring forces of all tether segements.

KiteModels.lift_drag — Function

lift_drag(s::AKM)

Return a tuple of the scalar lift and drag forces.

Example:

lift, drag = lift_drag(s)

KiteModels.lift_over_drag — Function

lift_over_drag(s::AKM)

Return the lift-over-drag ratio.

KiteModels.v_wind_kite — Function

v_wind_kite(s::AKM)

Return the vector of the wind speed at the height of the kite.

High level simulation interface

KiteModels.init_sim — Function

init_sim(s; t_end=1.0, stiffness_factor=0.035, prn=false)

Initialises the integrator of the model.

Parameters:

s: an instance of an abstract kite model

t_end: end time of the simulation; normally not needed

stiffness_factor: factor applied to the tether stiffness during initialisation

prn: if set to true, print the detailed solver results

Returns: An instance of a DAE integrator.

KiteModels.next_step — Function

next_step(s, integrator; v_ro = 0.0, v_wind_gnd=s.set.v_wind, wind_dir=0.0, dt=1/s.set.sample_freq)

Calculates the next simulation step.

Parameters:

s: an instance of an abstract kite model

integrator: an integrator instances as returned by the function @refinit_sim

v_ro: reel out speed in m/s

vwindgnd: wind speed at reference height in m/s

wind_dir: wind direction in radians

dt: time step in seconds

Only the first two parameters are required.

Returns: The end time of the time step in seconds.

Low level simulation interface

KiteModels.clear — Function

clear(s::KPS3)

Initialize the kite power model.

KiteModels.find_steady_state — Function

find_steady_state(s::KPS4; prn=false, delta = 0.0, stiffness_factor=0.035)

Find an initial equilibrium, based on the inital parameters l_tether, elevation and v_reel_out.

KiteModels.residual! — Function

residual!(res, yd, y::MVector{S, SimFloat}, s::KPS3, time) where S

N-point tether model, one point kite at the top:
Inputs:
State vector y   = pos1,  pos2, ... , posn,  vel1,  vel2, . .., veln,  length, v_reel_out
Derivative   yd  = posd1, posd2, ..., posdn, veld1, veld2, ..., veldn, lengthd, v_reel_outd
Output:
Residual     res = res1, res2 = vel1-posd1,  ..., veld1-acc1, ...

Additional parameters:
s: Struct with work variables, type KPS3
S: The dimension of the state vector

The number of the point masses of the model N = (S-2)/6, the state of each point is represented by two 3 element vectors.

residual!(res, yd, y::MVector{S, SimFloat}, s::KPS3, time) where S

N-point tether model, one point kite at the top:
Inputs:
State vector y   = pos1, pos2, ..., posn, vel1, vel2, ..., veln
Derivative   yd  = vel1, vel2, ..., veln, acc1, acc2, ..., accn
Output:
Residual     res = res1, res2 = pos1,  ..., vel1, ...

Additional parameters:
s: Struct with work variables, type KPS3
S: The dimension of the state vector

The number of the point masses of the model N = S/6, the state of each point is represented by two 3 element vectors.

Environment

KiteModels.calc_rho — Function

calc_rho(s, height)

Calculate the air densisity as function of height.

KiteModels.calc_wind_factor — Function

calc_wind_factor(s, height, profile_law=s.set.profile_law)

Calculate the relative wind speed at a given height and reference height.

Helper functions

KiteModels.calc_drag — Function

calc_drag(s::KPS3, v_segment, unit_vector, rho, last_tether_drag, v_app_perp)

Calculate the drag of one tether segment, result stored in parameter last_tether_drag. Return the norm of the apparent wind velocity.

KiteModels.calc_set_cl_cd — Function

calc_set_cl_cd(s, vec_c, v_app)

Calculate the lift over drag ratio as a function of the direction vector of the last tether segment, the current depower setting and the apparent wind speed. Set the calculated CL and CD values in the struct s.

KiteModels.calc_aero_forces — Function

posB, posC, posD: position of the kite particles B, C, and D vB, vC, vD: velocity of the kite particles B, C, and D rho: air density [kg/m^3] reldepower: value between 0.0 and 1.0 relsteering: value between -1.0 and +1.0

KiteModels.calc_particle_forces — Function

Calculate the drag force of the tether segment, defined by the parameters pos1, pos2, vel1 and vel2 and distribute it equally on the two particles, that are attached to the segment. The result is stored in the array s.forces.

KiteModels.initial_kite_ref_frame — Function

Calculate the initial orientation of the kite based on the last tether segment and the apparent wind speed.

Parameters: vecc: (posn-2) - (pos_n-1) n: number of particles without the three kite particles that do not belong to the main thether (P1, P2 and P3). Returns: x, y, z: the unit vectors of the kite reference frame in the ENU reference frame

KiteModels.inner_loop — Function

Calculate the forces, acting on all particles. vwindtether: out parameter forces: out parameter

KiteModels.loop — Function

Calculate the vector res1 and calculate res2 using loops that iterate over all tether segments.

KiteModels.find_steady_state_inner — Function

find_steady_state(s::KPS3, prn=false)

Find an initial equilibrium, based on the inital parameters l_tether, elevation and v_reel_out.

KiteModels.get_particles — Function

Calculate the initial positions of the particels representing a 4-point kite, connected to a kite control unit (KCU). Parameters:

mk: relative nose distance