Exported Types

Basic types

KiteModels.SimFloat — Type

const SimFloat = Float64

This type is used for all real variables, used in the Simulation. Possible alternatives: Float32, Double64, Dual Other types than Float64 or Float32 do require support of Julia types by the solver.

KiteModels.KVec3 — Type

const KVec3 = MVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, mutable.

KiteModels.SVec3 — Type

const SVec3 = SVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, immutable.

KiteModels.AbstractKiteModel — Type

abstract type AbstractKiteModel

All kite models must inherit from this type. All methods that are defined on this type must work with all kite models. All exported methods must work on this type.

KiteModels.AKM — Type

const AKM = AbstractKiteModel

Short alias for the AbstractKiteModel.

Struct KPS3 and KPS4 and RamAirKite

KiteModels.KPS3 — Type

mutable struct KPS3{S, T, P} <: AbstractKiteModel

State of the kite power system. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

T: Vector type, e.g. MVector{3, SimFloat}

P: number of points of the system, segments+1

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

kcu::KCU: Reference to the KCU model (Kite Control Unit as implemented in the package KitePodModels

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

wm::Union{Nothing, AbstractWinchModel}: Reference to winch model as implemented in the package WinchModels Default: nothing

iter::Int64: Iterations, number of calls to the function residual! Default: 0

calc_cl::Dierckx.Spline1D: Function for calculation the lift coefficent, using a spline based on the provided value pairs.

calc_cd::Dierckx.Spline1D: Function for calculation the drag coefficent, using a spline based on the provided value pairs.

v_wind::Any: wind vector at the height of the kite Default: zeros(S, 3)

v_wind_gnd::Any: wind vector at reference height Default: zeros(S, 3)

v_wind_tether::Any: wind vector used for the calculation of the tether drag Default: zeros(S, 3)

v_apparent::Any: apparent wind vector at the kite Default: zeros(S, 3)

v_app_perp::Any: vector, perpendicular to vapparent; output of calcdrag Default: zeros(S, 3)

alpha_2::Any: angle of attack of the kite; output of setclcd! Default: 0.0

drag_force::Any: drag force of kite and bridle; output of calcaeroforces Default: zeros(S, 3)

lift_force::Any: lift force of the kite; output of calcaeroforces Default: zeros(S, 3)

steering_force::Any: steering force acting on the kite; output of calcaeroforces Default: zeros(S, 3)

last_force::Any: Default: zeros(S, 3)

spring_force::Any: spring force of the current tether segment, output of calc_res Default: zeros(S, 3)

total_forces::Any: Default: zeros(S, 3)

force::Any: sum of spring and drag forces acting on the current segment, output of calc_res Default: zeros(S, 3)

unit_vector::Any: unit vector in the direction of the current tether segment, output of calc_res Default: zeros(S, 3)

av_vel::Any: average velocity of the current tether segment, output of calc_res Default: zeros(S, 3)

kite_y::Any: y-vector of the kite fixed referense frame, output of calcaeroforces Default: zeros(S, 3)

segment::Any: vector representing one tether segment (p1-p2) Default: zeros(S, 3)

last_tether_drag::Any: vector of the drag force of the last calculated tether segment Default: zeros(S, 3)

vec_z::Any: z vector of the kite reference frame Default: zeros(S, 3)

res1::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: part one of the residual, difference between pos' and vel, non-flat, mainly for unit testing Default: zeros(SVector{P, KVec3})

res2::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: part two of the residual, difference between vel' and acc, non-flat, mainly for unit testing Default: zeros(SVector{P, KVec3})

pos::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the positions of the particles Default: zeros(SVector{P, KVec3})

vel_kite::Any: velocity vector of the kite Default: zeros(S, 3)

seg_area::Any: area of one tether segment [m²] Default: zero(S)

bridle_area::Any: total actual bridle area [m²] Default: zero(S)

c_spring::Any: spring constant, depending on the length of the tether segment Default: zero(S)

segment_length::Any: unstressed segment length [m] Default: 0.0

damping::Any: damping factor, depending on the length of the tether segment Default: zero(S)

last_v_app_norm_tether::Any: last norm of the apparent wind speed at a tether segment [m/s] Default: zero(S)

param_cl::Any: lift coefficient of the kite, depending on the angle of attack Default: 0.2

param_cd::Any: drag coefficient of the kite, depending on the angle of attack Default: 1.0

cor_steering::Any: correction term for the steering, in paper named i_(s,c), Eq. 30 Default: zero(S)

psi::Any: azimuth angle in radian; inital value is zero Default: zero(S)

beta::Any: elevation angle in radian; initial value about 70 degrees Default: deg2rad((se()).elevation)

alpha_depower::Any: depower angle [deg] Default: 0.0

t_0::Any: relative start time of the current time interval Default: 0.0

v_reel_out::Any: reel out speed of the winch [m/s] Default: 0.0

last_v_reel_out::Any: reel out speed during the last time step Default: 0.0

l_tether::Any: unstretched tether length Default: 0.0

rho::Any: air density at the height of the kite Default: 0.0

depower::Any: actual relative depower setting, must be between 0 .. 1.0 Default: 0.0

steering::Any: actual relative steering setting, must be between -1.0 .. 1.0 Default: 0.0

kcu_steering::Any: steering after the kcu, before applying offset and depower sensitivity, -1.0 .. 1.0 Default: 0.0

stiffness_factor::Any: factor for the tether stiffness, used to find the steady state with a low stiffness first Default: 1.0

initial_masses::StaticArraysCore.MVector{P, S} where {S, P}: initial masses of the point masses Default: ones(P)

masses::StaticArraysCore.MVector{P, S} where {S, P}: current masses, depending on the total tether length Default: ones(P)

forces::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the forces, acting on the particles Default: zeros(SVector{P, KVec3})

sync_speed::Union{Nothing, S} where S: synchronous speed of the motor/ generator Default: 0.0

set_torque::Union{Nothing, S} where S: set_torque of the motor/generator Default: nothing

KiteModels.KPS4 — Type

mutable struct KPS4{S, T, P, Q, SP} <: AbstractKiteModel

State of the kite power system, using a 4 point kite model. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

T: Vector type, e.g. MVector{3, SimFloat}

P: number of points of the system, segments+1

Q: number of springs in the system, P-1

SP: struct type, describing a spring

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

kcu::KCU: Reference to the KCU model (Kite Control Unit as implemented in the package KitePodModels

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

wm::AbstractWinchModel: Reference to winch model as implemented in the package WinchModels

iter::Int64: Iterations, number of calls to the function residual! Default: 0

calc_cl::Dierckx.Spline1D: Function for calculation the lift coefficient, using a spline based on the provided value pairs.

calc_cd::Dierckx.Spline1D: Function for calculation the drag coefficient, using a spline based on the provided value pairs.

v_wind::Any: wind vector at the height of the kite Default: zeros(S, 3)

v_wind_gnd::Any: wind vector at reference height Default: zeros(S, 3)

v_wind_tether::Any: wind vector used for the calculation of the tether drag Default: zeros(S, 3)

v_apparent::Any: apparent wind vector at the kite Default: zeros(S, 3)

bridle_factor::Any: bridlefactor = set.lbridle/bridle_length(set) Default: 1.0

side_cl::Any: side lift coefficient, the difference of the left and right lift coefficients Default: 0.0

drag_force::Any: drag force of kite and bridle; output of calcaeroforces! Default: zeros(S, 3)

side_force::Any: side_force acting on the kite Default: zeros(S, 3)

f_d::Any: Default: zeros(S, 3)

f_s::Any: Default: zeros(S, 3)

ks::Any: max_steering angle in radian Default: 0.0

lift_force::Any: lift force of the kite; output of calcaeroforces! Default: zeros(S, 3)

spring_force::Any: spring force of the current tether segment, output of calcparticleforces! Default: zeros(S, 3)

last_force::Any: last winch force Default: zeros(S, 3)

res1::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the residual one (pos,vel) for debugging and unit tests Default: zeros(SVector{P, KVec3})

res2::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the residual two (vel,acc) for debugging and unit tests Default: zeros(SVector{P, KVec3})

pos::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the actual positions as output for the user Default: zeros(SVector{P, KVec3})

vel::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the actual velocities as output for the user Default: zeros(SVector{P, KVec3})

vel_kite::Any: velocity vector of the kite Default: zeros(S, 3)

segment_length::Any: unstressed segment length [m] Default: 0.0

param_cl::Any: lift coefficient of the kite, depending on the angle of attack Default: 0.2

param_cd::Any: drag coefficient of the kite, depending on the angle of attack Default: 1.0

psi::Any: azimuth angle in radian; initial value is zero Default: zero(S)

alpha_depower::Any: depower angle [deg] Default: 0.0

pitch::Any: pitch angle [rad] Default: 0.0

pitch_rate::Any: pitch rate [rad/s] Default: 0.0

alpha_2::Any: aoa at particle B Default: 0.0

alpha_3::Any: aoa at particle C Default: 0.0

alpha_4::Any: aoa at particle D Default: 0.0

side_slip::Any: side_slip angle [rad] Default: 0.0

t_0::Any: relative start time of the current time interval Default: 0.0

v_reel_out::Any: reel out speed of the winch Default: 0.0

last_v_reel_out::Any: reel out speed at the last time step Default: 0.0

l_tether::Any: unstretched tether length Default: 0.0

rho::Any: air density at the height of the kite Default: 0.0

depower::Any: actual relative depower setting, must be between 0 .. 1.0 Default: 0.0

steering::Any: actual relative steering setting, must be between -1.0 .. 1.0 Default: 0.0

kcu_steering::Any: steering after the kcu, before applying offset and depower sensitivity, -1.0 .. 1.0 Default: 0.0

stiffness_factor::Any: multiplier for the stiffniss of tether and bridle Default: 1.0

initial_masses::StaticArraysCore.MVector{P, S} where {S, P}: initial masses of the point masses Default: ones(P)

masses::StaticArraysCore.MVector{P, S} where {S, P}: current masses, depending on the total tether length Default: zeros(P)

springs::StaticArraysCore.MVector: vector of the springs, defined as struct Default: zeros(SP, Q)

forces::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the forces, acting on the particles Default: zeros(SVector{P, KVec3})

sync_speed::Union{Nothing, S} where S: synchronous speed of the motor/ generator Default: 0.0

set_torque::Union{Nothing, S} where S: set_torque of the motor/generator Default: nothing

set_force::Union{Nothing, S} where S: set value of the force at the winch, for logging only Default: nothing

bearing::Union{Nothing, S} where S: set value of the bearing angle in radians, for logging only Default: nothing

attractor::Union{Nothing, StaticArraysCore.SVector{2, S}} where S: coordinates of the attractor point [azimuth, elevation] in radian, for logging only Default: nothing

x::Any: x vector of kite reference frame Default: zeros(S, 3)

y::Any: y vector of kite reference frame Default: zeros(S, 3)

z::Any: z vector of kite reference frame Default: zeros(S, 3)

KiteModels.RamAirKite — Type

mutable struct RamAirKite{S, V, P} <: AbstractKiteModel

State of the kite power system, using a quaternion kite model and three steering lines to the ground. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

V: Vector type, e.g. KVec3

P: number of tether points of the system, 3segments+3

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

wing::RamAirWing: Reference to the geometric wing model

aero::BodyAerodynamics: Reference to the aerodynamic wing model

vsm_solver::Solver: Reference to the VSM aerodynamics solver

point_system::PointMassSystem: Reference to the point mass system with points, segments, pulleys and tethers

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

pos::Matrix: tether positions Default: zeros(S, 3, P)

segment_lengths::Any: unstressed segment lengths of the three tethers [m] Default: zeros(S, 3)

t_0::Any: relative start time of the current time interval Default: 0.0

tether_lengths::Any: unstretched tether length Default: zeros(S, 3)

rho::Any: air density at the height of the kite Default: 0.0

masses::Any: tether masses Default: zeros(S, P)

c_spring::Any: unit spring coefficient Default: zeros(S, 3)

damping::Any: unit damping coefficient Default: zeros(S, 3)

torque_control::Bool: whether or not to use torque control instead of speed control Default: false

e_x::Any: x vector of kite reference frame Default: zeros(S, 3)

e_y::Any: y vector of kite reference frame Default: zeros(S, 3)

e_z::Any: z vector of kite reference frame Default: zeros(S, 3)

sys::Union{Nothing, ModelingToolkit.ODESystem}: Simplified system of the mtk model Default: nothing

vel_kite::Any: Velocity of the kite Default: zeros(S, 3)

I_b::Any: Inertia around kite x y and z axis of the body frame Default: zeros(S, 3)

u0map::Union{Nothing, Array{Pair{Symbolics.Num, S}, 1}} where S: Initialization values for kite state Default: nothing

p0map::Union{Nothing, Array{Pair{Symbolics.Num, S}, 1}} where S: Initialization values for kite parameters Default: nothing

bridle_fracs::Any: X coordinate on normalized 2d foil of bridle attachments Default: [0.088, 0.31, 0.58, 0.93]

crease_frac::Any: Default: 0.82

top_bridle_points::Vector: The top bridle points that are not on the kite, in CAD frame Default: [[0.290199, 0.784697, -2.61305], [0.392683, 0.785271, -2.61201], [0.498202, 0.786175, -2.62148], [0.535543, 0.786175, -2.62148]]

bridle_tether_diameter::Float64: Tether diameter of tethers in bridle system [mm] Default: 2.0

power_tether_diameter::Float64: Tether diameter of the power tethers [mm] Default: 2.0

steering_tether_diameter::Float64: Tether diameter of the steering tethers [mm] Default: 1.0

iter::Int64: Number of solve! calls Default: 0

unknowns_vec::Vector{Float64}: Default: zeros(SimFloat, 3)

set_set_values::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:292 =# nothing end

set_measure::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:293 =# nothing end

set_vsm::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:294 =# nothing end

set_unknowns::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:295 =# nothing end

get_state::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:297 =# nothing end

get_y::Function: Default: ()->begin #= /home/runner/work/KiteModels.jl/KiteModels.jl/src/ramairkite.jl:298 =# nothing end

prob::Union{Nothing, SciMLBase.ODEProblem}: Default: nothing

init_prob::Union{Nothing, SciMLBase.ODEProblem}: Default: nothing

integrator::Union{Nothing, OrdinaryDiffEqCore.ODEIntegrator, Sundials.CVODEIntegrator}: Default: nothing

init_integrator::Union{Nothing, OrdinaryDiffEqCore.ODEIntegrator, Sundials.CVODEIntegrator}: Default: nothing

These structs store the state of the one point model and four point model. Only in unit tests it is allowed to access the members directly, otherwise use the input and output functions.