Exported Types

Basic types

KiteModels.SimFloat — Type

const SimFloat = Float64

This type is used for all real variables, used in the Simulation. Possible alternatives: Float32, Double64, Dual Other types than Float64 or Float32 do require support of Julia types by the solver.

KiteModels.KVec3 — Type

const KVec3 = MVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, mutable.

KiteModels.SVec3 — Type

const SVec3 = SVector{3, SimFloat}

Basic 3-dimensional vector, stack allocated, immutable.

KiteModels.AbstractKiteModel — Type

abstract type AbstractKiteModel

All kite models must inherit from this type. All methods that are defined on this type must work with all kite models. All exported methods must work on this type.

KiteModels.AKM — Type

const AKM = AbstractKiteModel

Short alias for the AbstractKiteModel.

Struct KPS3 and KPS4 and KPS4_3L

KiteModels.KPS3 — Type

mutable struct KPS3{S, T, P} <: AbstractKiteModel

State of the kite power system. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

T: Vector type, e.g. MVector{3, SimFloat}

P: number of points of the system, segments+1

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

kcu::KCU: Reference to the KCU model (Kite Control Unit as implemented in the package KitePodModels

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

wm::Union{Nothing, AbstractWinchModel}: Reference to winch model as implemented in the package WinchModels Default: nothing

iter::Int64: Iterations, number of calls to the function residual! Default: 0

calc_cl::Dierckx.Spline1D: Function for calculation the lift coefficent, using a spline based on the provided value pairs.

calc_cd::Dierckx.Spline1D: Function for calculation the drag coefficent, using a spline based on the provided value pairs.

v_wind::Any: wind vector at the height of the kite Default: zeros(S, 3)

v_wind_gnd::Any: wind vector at reference height Default: zeros(S, 3)

v_wind_tether::Any: wind vector used for the calculation of the tether drag Default: zeros(S, 3)

v_apparent::Any: apparent wind vector at the kite Default: zeros(S, 3)

v_app_perp::Any: vector, perpendicular to vapparent; output of calcdrag Default: zeros(S, 3)

drag_force::Any: drag force of kite and bridle; output of calcaeroforces Default: zeros(S, 3)

lift_force::Any: lift force of the kite; output of calcaeroforces Default: zeros(S, 3)

steering_force::Any: steering force acting on the kite; output of calcaeroforces Default: zeros(S, 3)

last_force::Any: Default: zeros(S, 3)

spring_force::Any: spring force of the current tether segment, output of calc_res Default: zeros(S, 3)

total_forces::Any: Default: zeros(S, 3)

force::Any: sum of spring and drag forces acting on the current segment, output of calc_res Default: zeros(S, 3)

unit_vector::Any: unit vector in the direction of the current tether segment, output of calc_res Default: zeros(S, 3)

av_vel::Any: average velocity of the current tether segment, output of calc_res Default: zeros(S, 3)

kite_y::Any: y-vector of the kite fixed referense frame, output of calcaeroforces Default: zeros(S, 3)

segment::Any: vector representing one tether segment (p1-p2) Default: zeros(S, 3)

last_tether_drag::Any: vector of the drag force of the last calculated tether segment Default: zeros(S, 3)

vec_z::Any: z vector of the kite reference frame Default: zeros(S, 3)

res1::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: part one of the residual, difference between pos' and vel, non-flat, mainly for unit testing Default: zeros(SVector{P, KVec3})

res2::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: part two of the residual, difference between vel' and acc, non-flat, mainly for unit testing Default: zeros(SVector{P, KVec3})

pos::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the positions of the particles Default: zeros(SVector{P, KVec3})

vel_kite::Any: velocity vector of the kite Default: zeros(S, 3)

seg_area::Any: area of one tether segment [m²] Default: zero(S)

bridle_area::Any: total actual bridle area [m²] Default: zero(S)

c_spring::Any: spring constant, depending on the length of the tether segment Default: zero(S)

segment_length::Any: unstressed segment length [m] Default: 0.0

damping::Any: damping factor, depending on the length of the tether segment Default: zero(S)

last_v_app_norm_tether::Any: last norm of the apparent wind speed at a tether segment [m/s] Default: zero(S)

param_cl::Any: lift coefficient of the kite, depending on the angle of attack Default: 0.2

param_cd::Any: drag coefficient of the kite, depending on the angle of attack Default: 1.0

cor_steering::Any: correction term for the steering, in paper named i_(s,c), Eq. 30 Default: zero(S)

psi::Any: azimuth angle in radian; inital value is zero Default: zero(S)

beta::Any: elevation angle in radian; initial value about 70 degrees Default: deg2rad((se()).elevation)

alpha_depower::Any: depower angle [deg] Default: 0.0

t_0::Any: relative start time of the current time interval Default: 0.0

v_reel_out::Any: reel out speed of the winch [m/s] Default: 0.0

last_v_reel_out::Any: reel out speed during the last time step Default: 0.0

l_tether::Any: unstretched tether length Default: 0.0

rho::Any: air density at the height of the kite Default: 0.0

depower::Any: actual relative depower setting, must be between 0 .. 1.0 Default: 0.0

steering::Any: actual relative steering setting, must be between -1.0 .. 1.0 Default: 0.0

stiffness_factor::Any: factor for the tether stiffness, used to find the steady state with a low stiffness first Default: 1.0

initial_masses::StaticArraysCore.MVector{P, S} where {S, P}: initial masses of the point masses Default: ones(P)

masses::StaticArraysCore.MVector{P, S} where {S, P}: current masses, depending on the total tether length Default: ones(P)

forces::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the forces, acting on the particles Default: zeros(SVector{P, KVec3})

sync_speed::Union{Nothing, S} where S: synchronous speed of the motor/ generator Default: 0.0

set_torque::Union{Nothing, S} where S: set_torque of the motor/generator Default: nothing

KiteModels.KPS4 — Type

mutable struct KPS4{S, T, P, Q, SP} <: AbstractKiteModel

State of the kite power system, using a 4 point kite model. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

T: Vector type, e.g. MVector{3, SimFloat}

P: number of points of the system, segments+1

Q: number of springs in the system, P-1

SP: struct type, describing a spring

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

kcu::KCU: Reference to the KCU model (Kite Control Unit as implemented in the package KitePodModels

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

wm::AbstractWinchModel: Reference to winch model as implemented in the package WinchModels

iter::Int64: Iterations, number of calls to the function residual! Default: 0

calc_cl::Dierckx.Spline1D: Function for calculation the lift coefficent, using a spline based on the provided value pairs.

calc_cd::Dierckx.Spline1D: Function for calculation the drag coefficent, using a spline based on the provided value pairs.

v_wind::Any: wind vector at the height of the kite Default: zeros(S, 3)

v_wind_gnd::Any: wind vector at reference height Default: zeros(S, 3)

v_wind_tether::Any: wind vector used for the calculation of the tether drag Default: zeros(S, 3)

v_apparent::Any: apparent wind vector at the kite Default: zeros(S, 3)

bridle_factor::Any: bridlefactor = set.lbridle/bridle_length(set) Default: 1.0

drag_force::Any: drag force of kite and bridle; output of calcaeroforces! Default: zeros(S, 3)

lift_force::Any: lift force of the kite; output of calcaeroforces! Default: zeros(S, 3)

spring_force::Any: spring force of the current tether segment, output of calcparticleforces! Default: zeros(S, 3)

last_force::Any: last winch force Default: zeros(S, 3)

res1::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the residual one (pos,vel) for debugging and unit tests Default: zeros(SVector{P, KVec3})

res2::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the residual two (vel,acc) for debugging and unit tests Default: zeros(SVector{P, KVec3})

pos::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: a copy of the actual positions as output for the user Default: zeros(SVector{P, KVec3})

vel_kite::Any: velocity vector of the kite Default: zeros(S, 3)

segment_length::Any: unstressed segment length [m] Default: 0.0

param_cl::Any: lift coefficient of the kite, depending on the angle of attack Default: 0.2

param_cd::Any: drag coefficient of the kite, depending on the angle of attack Default: 1.0

psi::Any: azimuth angle in radian; inital value is zero Default: zero(S)

alpha_depower::Any: depower angle [deg] Default: 0.0

alpha_2::Any: aoa at paricle B Default: 0.0

alpha_2b::Any: aoa at paricle B, corrected formula Default: 0.0

alpha_3::Any: aoa at paricle C Default: 0.0

alpha_3b::Any: Default: 0.0

alpha_4::Any: aoa at paricle D Default: 0.0

alpha_4b::Any: Default: 0.0

t_0::Any: relative start time of the current time interval Default: 0.0

v_reel_out::Any: reel out speed of the winch Default: 0.0

last_v_reel_out::Any: reel out speed at the last time step Default: 0.0

l_tether::Any: unstretched tether length Default: 0.0

rho::Any: air density at the height of the kite Default: 0.0

depower::Any: actual relative depower setting, must be between 0 .. 1.0 Default: 0.0

steering::Any: actual relative steering setting, must be between -1.0 .. 1.0 Default: 0.0

stiffness_factor::Any: multiplier for the stiffniss of tether and bridle Default: 1.0

initial_masses::StaticArraysCore.MVector{P, S} where {S, P}: initial masses of the point masses Default: ones(P)

masses::StaticArraysCore.MVector{P, S} where {S, P}: current masses, depending on the total tether length Default: zeros(P)

springs::StaticArraysCore.MVector: vector of the springs, defined as struct Default: zeros(SP, Q)

forces::StaticArraysCore.SVector{P, StaticArraysCore.MVector{3, Float64}} where P: vector of the forces, acting on the particles Default: zeros(SVector{P, KVec3})

sync_speed::Union{Nothing, S} where S: synchronous speed of the motor/ generator Default: 0.0

set_torque::Union{Nothing, S} where S: set_torque of the motor/generator Default: nothing

x::Any: x vector of kite reference frame Default: zeros(S, 3)

y::Any: y vector of kite reference frame Default: zeros(S, 3)

z::Any: z vector of kite reference frame Default: zeros(S, 3)

KiteModels.KPS4_3L — Type

mutable struct KPS4_3L{S, T, P, Q, SP} <: AbstractKiteModel

State of the kite power system, using a 3 point kite model and three steering lines to the ground. Parameters:

S: Scalar type, e.g. SimFloat In the documentation mentioned as Any, but when used in this module it is always SimFloat and not Any.

T: Vector type, e.g. MVector{3, SimFloat}

P: number of points of the system, segments+3

Q: number of springs in the system, P-1

SP: struct type, describing a spring

Normally a user of this package will not have to access any of the members of this type directly, use the input and output functions instead.

set::Settings: Reference to the settings struct

am::AtmosphericModel: Reference to the atmospheric model as implemented in the package AtmosphericModels Default: AtmosphericModel()

calc_cl::Any: Function for calculation the lift coefficent, using a spline based on the provided value pairs. Default: Spline1D((se()).alphacl, (se()).cllist)

calc_cd::Any: Function for calculation the drag coefficent, using a spline based on the provided value pairs. Default: Spline1D((se()).alphacd, (se()).cdlist)

motors::StaticArraysCore.SVector{3, AbstractWinchModel}: Reference to the motor models as implemented in the package WinchModels. index 1: middle motor, index 2: left motor, index 3: right motor

iter::Int64: Iterations, number of calls to the function residual! Default: 0

v_wind::Any: wind vector at the height of the kite Default: zeros(S, 3)

v_wind_gnd::Any: wind vector at reference height Default: zeros(S, 3)

v_wind_tether::Any: wind vector used for the calculation of the tether drag Default: zeros(S, 3)

v_apparent::Any: apparent wind vector at the kite Default: zeros(S, 3)

drag_force::Any: drag force of kite and bridle; output of calcaeroforces! Default: zeros(S, 3)

lift_force::Any: lift force of the kite; output of calcaeroforces! Default: zeros(S, 3)

spring_force::Any: spring force of the current tether segment, output of calcparticleforces! Default: zeros(S, 3)

winch_forces::StaticArraysCore.SVector{3}: last winch force Default: [zeros(S, 3) for _ = 1:3]

res1::StaticArraysCore.SVector{P, T} where {T, P}: a copy of the residual one (pos,vel) for debugging and unit tests Default: zeros(SVector{P, T})

res2::StaticArraysCore.SVector{P, T} where {T, P}: a copy of the residual two (vel,acc) for debugging and unit tests Default: zeros(SVector{P, T})

pos::StaticArraysCore.SVector{P, T} where {T, P}: a copy of the actual positions as output for the user Default: zeros(SVector{P, T})

stable_pos::StaticArraysCore.SVector{P, T} where {T, P}: Default: zeros(SVector{P, T})

vel::StaticArraysCore.SVector{P, T} where {T, P}: Default: zeros(SVector{P, T})

posd::StaticArraysCore.SVector{P, T} where {T, P}: Default: zeros(SVector{P, T})

veld::StaticArraysCore.SVector{P, T} where {T, P}: Default: zeros(SVector{P, T})

vel_kite::Any: velocity vector of the kite Default: zeros(S, 3)

steering_vel::Any: Default: zeros(S, 3)

segment_lengths::Any: unstressed segment lengths of the three tethers [m] Default: zeros(S, 3)

param_cl::Any: lift coefficient of the kite, depending on the angle of attack Default: 0.2

param_cd::Any: drag coefficient of the kite, depending on the angle of attack Default: 1.0

psi::Any: azimuth angle in radian; inital value is zero Default: zero(S)

t_0::Any: relative start time of the current time interval Default: 0.0

reel_out_speeds::Any: reel out speed of the winch Default: zeros(S, 3)

tether_lengths::Any: unstretched tether length Default: zeros(S, 3)

steering_pos::StaticArraysCore.MVector{2}: lengths of the connections of the steering tethers to the kite Default: zeros(S, 2)

rho::Any: air density at the height of the kite Default: 0.0

stiffness_factor::Any: multiplier for the stiffniss of tether and bridle Default: 1.0

initial_masses::StaticArraysCore.MVector{P, S} where {S, P}: initial masses of the point masses Default: ones(P)

masses::StaticArraysCore.MVector{P, S} where {S, P}: current masses, depending on the total tether length Default: zeros(P)

springs::StaticArraysCore.MVector: vector of the springs, defined as struct Default: zeros(SP, Q)

forces::StaticArraysCore.SVector{P, T} where {T, P}: vector of the forces, acting on the particles Default: zeros(SVector{P, T})

set_values::StaticArraysCore.MVector{3, Float64}: synchronous speed or torque of the motor/ generator Default: zeros(KVec3)

torque_control::Bool: Default: false

e_x::Any: x vector of kite reference frame Default: zeros(S, 3)

e_y::Any: y vector of kite reference frame Default: zeros(S, 3)

e_z::Any: z vector of kite reference frame Default: zeros(S, 3)

e_r::Any: Default: zeros(S, 3)

num_E::Int64: Point number of E Default: 0

num_C::Int64: Point number of C Default: 0

num_D::Int64: Point number of D Default: 0

mtk::Bool: mtk variables Default: false

set_values_idx::Union{Nothing, ModelingToolkit.ParameterIndex}: Default: nothing

v_wind_gnd_idx::Union{Nothing, ModelingToolkit.ParameterIndex}: Default: nothing

stiffness_factor_idx::Union{Nothing, ModelingToolkit.ParameterIndex}: Default: nothing

v_wind_idx::Union{Nothing, ModelingToolkit.ParameterIndex}: Default: nothing

prob::Union{Nothing, SciMLBase.ODEProblem}: Default: nothing

get_pos::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_steering_pos::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_line_acc::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_kite_vel::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_winch_forces::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_tether_lengths::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_tether_speeds::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_L_C::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_L_D::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_D_C::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

get_D_D::Union{Nothing, SymbolicIndexingInterface.MultipleGetters, SymbolicIndexingInterface.TimeDependentObservedFunction}: Default: nothing

half_drag_force::StaticArraysCore.SVector{P, T} where {T, P}: Default: zeros(SVector{P, T})

num_A::Int64: residual variables Default: 0

L_C::Any: Default: zeros(S, 3)

L_D::Any: Default: zeros(S, 3)

D_C::Any: Default: zeros(S, 3)

D_D::Any: Default: zeros(S, 3)

F_steering_c::Any: Default: zeros(S, 3)

F_steering_d::Any: Default: zeros(S, 3)

dL_dα::Any: Default: zeros(S, 3)

dD_dα::Any: Default: zeros(S, 3)

v_cx::Any: Default: zeros(S, 3)

v_dx::Any: Default: zeros(S, 3)

v_dy::Any: Default: zeros(S, 3)

v_dz::Any: Default: zeros(S, 3)

v_cy::Any: Default: zeros(S, 3)

v_cz::Any: Default: zeros(S, 3)

v_kite::Any: Default: zeros(S, 3)

v_a::Any: Default: zeros(S, 3)

e_drift::Any: Default: zeros(S, 3)

v_a_xr::Any: Default: zeros(S, 3)

E_c::Any: Default: zeros(S, 3)

F::Any: Default: zeros(S, 3)

y_lc::Any: Default: 0.0

y_ld::Any: Default: 0.0

δ_left::Any: Default: 0.0

δ_right::Any: Default: 0.0

α_l::Any: Default: 0.0

α_r::Any: Default: 0.0

distance_c_e::Any: Default: 0

These structs store the state of the one point model and four point model. Only in unit tests it is allowed to access the members directly, otherwise use the input and output functions. The model KPS4_3L is not yet implemented, it is just a copy of KPS4.